Generic to convert objects to rmt_list

Generic to convert objects to rmt_list

rmt_as_list(...)

# S3 method for wel
rmt_as_list(wel, cwel, kper, ...)

# S3 method for drn
rmt_as_list(drn, cdrn, kper, ...)

# S3 method for chd
rmt_as_list(chd, cchd, kper, ...)

# S3 method for riv
rmt_as_list(riv, criv, kper, ...)

# S3 method for ghb
rmt_as_list(ghb, cghb, kper, ...)

Arguments

...
additional numeric vectors containing concentrations (one value for each feature) or column names. One vector/name for each species > 1 (in order).

wel
RMODFLOW wel object

cwel
concentrations in the wells for species 1. A value must be supplied for every well. Alternatively, a character can be entered referencing a column in the wel data set to use.

kper
numeric vector with the stress period numbers during which the list is active.

drn
RMODFLOW drn object

cdrn
concentrations in the drains for species 1. A value must be supplied for every drain. Alternatively, a character can be entered referencing a column in the drn data set to use.

chd
RMODFLOW chd object

cchd
concentrations in the constant-head cells for species 1. A value must be supplied for every constant-head cell. Alternatively, a character can be entered referencing a column in the chd data set to use.

riv
RMODFLOW riv object

criv
concentrations in the river cells for species 1. A value must be supplied for every river cell. Alternatively, a character can be entered referencing a column in the riv data set to use.

ghb
RMODFLOW ghb object

cghb
concentrations in the general-head boundary cells for species 1. A value must be supplied for every general-head boundary cell. Alternatively, a character can be entered referencing a column in the ghb data set to use.

Value

a rmt_list object

a rmt_list object

a rmt_list object

a rmt_list object

a rmt_list object

Details

kper should be one of the stress periods defined in the flow model unless only one stress period is present in the flow model.

Examples

library(RMODFLOW)

#> ! {RMODFLOW} is still in its experimental lifecycle stage.

#> 

#> ! Use at your own risk, and submit issues here:

#> 

#> ! <https://github.com/rogiersbart/RMODFLOW/issues>

#> 

m <- rmf_read(rmf_example_file('example-model.nam'), verbose = FALSE)
rmt_as_list(m$wel, 6.5, kper = 1)

#> RMODFLOW list with 1 features and 6 variables 
#> Active during stress period: 1 
#>   k i j   q IFACE parameter   name itype css1
#> 1 1 5 6 0.3     0     FALSE list_1     2  6.5


m$wel$data$chloride <- 6.5
rmt_as_list(m$wel, 'chloride', kper = 1)

#> RMODFLOW list with 1 features and 7 variables 
#> Active during stress period: 1 
#>   k i j   q IFACE parameter   name chloride itype css1
#> 1 1 5 6 0.3     0     FALSE list_1      6.5     2  6.5

# 3 species
cchd1 <- rnorm(nrow(m$chd$data), 70, 50)
rmt_as_list(m$chd, cchd1, kper = c(2, 3), cchd1 * 2, cchd1 / 2)

#> RMODFLOW list with 12 features and 9 variables 
#> Active during stress periods: 2 3 
#>    k  i  j shead ehead IFACE parameter     name itype          css1
#> 1  1  1 10  0.05  0.05     0      TRUE CHD_Par1     1  -0.002175836
#> 2  1  2 10  0.15  0.15     0      TRUE CHD_Par1     1  82.765852742
#> 3  1  3 10  0.25  0.25     0      TRUE CHD_Par1     1 -51.863180561
#> 4  1  4 10  0.35  0.35     0      TRUE CHD_Par1     1  69.721435663
#> 5  1  5 10  0.45  0.45     0      TRUE CHD_Par1     1 101.077636071
#> 6  1  1 10  0.95  0.95     0      TRUE CHD_Par2     1 127.420580301
#> 7  1  2 10  0.85  0.85     0      TRUE CHD_Par2     1 -21.090883049
#> 8  1  3 10  0.75  0.75     0      TRUE CHD_Par2     1  57.633734896
#> 9  1  4 10  0.65  0.65     0      TRUE CHD_Par2     1  57.790019661
#> 10 1  5 10  0.55  0.55     0      TRUE CHD_Par2     1  55.864727559
#> 11 1  1  1 30.00 30.00     0     FALSE   list_1     1  42.315030816
#> 12 1 10 10 29.00 29.00     0     FALSE   list_1     1 101.449102102
#>             css2          css3
#> 1  -4.351672e-03  -0.001087918
#> 2   1.655317e+02  41.382926371
#> 3  -1.037264e+02 -25.931590280
#> 4   1.394429e+02  34.860717831
#> 5   2.021553e+02  50.538818035
#> 6   2.548412e+02  63.710290151
#> 7  -4.218177e+01 -10.545441524
#> 8   1.152675e+02  28.816867448
#> 9   1.155800e+02  28.895009831
#> 10  1.117295e+02  27.932363780
#> 11  8.463006e+01  21.157515408
#> 12  2.028982e+02  50.724551051


# stress periods should be the same when flow object has more than 1 stress period
if (FALSE) {
rmt_as_list(m$chd, cchd1, kper = c(2, 3))
}