Create an RMT3DMS mt3dms object

rmt_create creates an RMT3DMS mt3dms object from rmt_package objects

rmt_create(
  ...,
  recreate_nam = FALSE,
  ftl,
  ftl_free = NULL,
  print = FALSE,
  basename = NULL,
  reset_timprs = FALSE
)

Arguments

...
(list of) RMT3DMS objects of class rmt_package to be included in the mt3dms object. If a nam object is not provided, it is added automatically.

recreate_nam
logical; if a nam object is supplied, should it be recreated from scratch ? Defaults to FALSE

ftl
path to the flow-transport link file in the nam object; typically '.ftl'. Only used when a nam object is not supplied or when recreate_nam = TRUE.

ftl_free
logical; is the flow-transport link file in the nam object written in free (formatted) format (TRUE) or binary (unformatted) (FALSE)? if NULL (default), it is guessed from reading ftl. Only used when a nam object is not supplied or when recreate_nam = TRUE.

basename
character specifying the basename of the files if the nam object is (re)created. The default (NULL) sets input basenames to 'input' and output to 'output'.

reset_timprs
logical, when a tob object is present, should the observed times be added to btn$timprs in order to write output at correct times? Defaults to FALSE. Note that this also resets btn$nprs

Value

a mt3dms object which is a list containing all MT3DMS/MT3D-USGS packages

See also

Examples

btn <- rmt_create_btn()
adv <- rmt_create_adv()
gcg <- rmt_create_gcg()
nam <- rmt_create_nam(btn, adv, gcg, ftl = 'output.ftl', ftl_free = TRUE)

rmt_create(btn, adv, gcg, nam)

#> $btn
#> $modflowstylearrays
#> [1] FALSE
#> 
#> $drycell
#> [1] FALSE
#> 
#> $legacy99storage
#> [1] FALSE
#> 
#> $ftlprint
#> [1] FALSE
#> 
#> $nowetdryprint
#> [1] FALSE
#> 
#> $omitdrycellbudget
#> [1] FALSE
#> 
#> $altwtsorb
#> [1] FALSE
#> 
#> $nlay
#> [1] 3
#> 
#> $nrow
#> [1] 10
#> 
#> $ncol
#> [1] 10
#> 
#> $nper
#> [1] 1
#> 
#> $ncomp
#> [1] 1
#> 
#> $mcomp
#> [1] 1
#> 
#> $tunit
#> [1] "s"
#> 
#> $lunit
#> [1] "m"
#> 
#> $munit
#> [1] "g"
#> 
#> $trnop
#>  [1]  TRUE  TRUE  TRUE  TRUE  TRUE FALSE FALSE FALSE FALSE FALSE
#> 
#> $laycon
#> [1] 1 1 1
#> 
#> $delr
#>  [1] 100 100 100 100 100 100 100 100 100 100
#> 
#> $delc
#>  [1] 100 100 100 100 100 100 100 100 100 100
#> 
#> $htop
#> RMODFLOW 2d array with 10 rows and 10 columns, representing the i & j dimensions. 
#> Not representing stress period data 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    0    0    0    0    0    0    0    0    0     0
#>  [2,]    0    0    0    0    0    0    0    0    0     0
#>  [3,]    0    0    0    0    0    0    0    0    0     0
#>  [4,]    0    0    0    0    0    0    0    0    0     0
#>  [5,]    0    0    0    0    0    0    0    0    0     0
#>  [6,]    0    0    0    0    0    0    0    0    0     0
#>  [7,]    0    0    0    0    0    0    0    0    0     0
#>  [8,]    0    0    0    0    0    0    0    0    0     0
#>  [9,]    0    0    0    0    0    0    0    0    0     0
#> [10,]    0    0    0    0    0    0    0    0    0     0
#> 
#> $dz
#> RMODFLOW 3d array with 10 rows, 10 columns and 3 layers, representing the i, j & k dimensions. 
#> Not representing stress period data 
#> , , 1
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]   10   10   10   10   10   10   10   10   10    10
#>  [2,]   10   10   10   10   10   10   10   10   10    10
#>  [3,]   10   10   10   10   10   10   10   10   10    10
#>  [4,]   10   10   10   10   10   10   10   10   10    10
#>  [5,]   10   10   10   10   10   10   10   10   10    10
#>  [6,]   10   10   10   10   10   10   10   10   10    10
#>  [7,]   10   10   10   10   10   10   10   10   10    10
#>  [8,]   10   10   10   10   10   10   10   10   10    10
#>  [9,]   10   10   10   10   10   10   10   10   10    10
#> [10,]   10   10   10   10   10   10   10   10   10    10
#> 
#> , , 2
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]   10   10   10   10   10   10   10   10   10    10
#>  [2,]   10   10   10   10   10   10   10   10   10    10
#>  [3,]   10   10   10   10   10   10   10   10   10    10
#>  [4,]   10   10   10   10   10   10   10   10   10    10
#>  [5,]   10   10   10   10   10   10   10   10   10    10
#>  [6,]   10   10   10   10   10   10   10   10   10    10
#>  [7,]   10   10   10   10   10   10   10   10   10    10
#>  [8,]   10   10   10   10   10   10   10   10   10    10
#>  [9,]   10   10   10   10   10   10   10   10   10    10
#> [10,]   10   10   10   10   10   10   10   10   10    10
#> 
#> , , 3
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]   10   10   10   10   10   10   10   10   10    10
#>  [2,]   10   10   10   10   10   10   10   10   10    10
#>  [3,]   10   10   10   10   10   10   10   10   10    10
#>  [4,]   10   10   10   10   10   10   10   10   10    10
#>  [5,]   10   10   10   10   10   10   10   10   10    10
#>  [6,]   10   10   10   10   10   10   10   10   10    10
#>  [7,]   10   10   10   10   10   10   10   10   10    10
#>  [8,]   10   10   10   10   10   10   10   10   10    10
#>  [9,]   10   10   10   10   10   10   10   10   10    10
#> [10,]   10   10   10   10   10   10   10   10   10    10
#> 
#> 
#> $prsity
#> RMODFLOW 3d array with 10 rows, 10 columns and 3 layers, representing the i, j & k dimensions. 
#> Not representing stress period data 
#> , , 1
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [2,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [3,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [4,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [5,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [6,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [7,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [8,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [9,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#> [10,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#> 
#> , , 2
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [2,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [3,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [4,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [5,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [6,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [7,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [8,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [9,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#> [10,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#> 
#> , , 3
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [2,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [3,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [4,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [5,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [6,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [7,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [8,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [9,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#> [10,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#> 
#> 
#> $icbund
#> RMODFLOW 3d array with 10 rows, 10 columns and 3 layers, representing the i, j & k dimensions. 
#> Not representing stress period data 
#> , , 1
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    1    1    1    1    1    1    1    1    1     1
#>  [2,]    1    1    1    1    1    1    1    1    1     1
#>  [3,]    1    1    1    1    1    1    1    1    1     1
#>  [4,]    1    1    1    1    1    1    1    1    1     1
#>  [5,]    1    1    1    1    1    1    1    1    1     1
#>  [6,]    1    1    1    1    1    1    1    1    1     1
#>  [7,]    1    1    1    1    1    1    1    1    1     1
#>  [8,]    1    1    1    1    1    1    1    1    1     1
#>  [9,]    1    1    1    1    1    1    1    1    1     1
#> [10,]    1    1    1    1    1    1    1    1    1     1
#> 
#> , , 2
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    1    1    1    1    1    1    1    1    1     1
#>  [2,]    1    1    1    1    1    1    1    1    1     1
#>  [3,]    1    1    1    1    1    1    1    1    1     1
#>  [4,]    1    1    1    1    1    1    1    1    1     1
#>  [5,]    1    1    1    1    1    1    1    1    1     1
#>  [6,]    1    1    1    1    1    1    1    1    1     1
#>  [7,]    1    1    1    1    1    1    1    1    1     1
#>  [8,]    1    1    1    1    1    1    1    1    1     1
#>  [9,]    1    1    1    1    1    1    1    1    1     1
#> [10,]    1    1    1    1    1    1    1    1    1     1
#> 
#> , , 3
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    1    1    1    1    1    1    1    1    1     1
#>  [2,]    1    1    1    1    1    1    1    1    1     1
#>  [3,]    1    1    1    1    1    1    1    1    1     1
#>  [4,]    1    1    1    1    1    1    1    1    1     1
#>  [5,]    1    1    1    1    1    1    1    1    1     1
#>  [6,]    1    1    1    1    1    1    1    1    1     1
#>  [7,]    1    1    1    1    1    1    1    1    1     1
#>  [8,]    1    1    1    1    1    1    1    1    1     1
#>  [9,]    1    1    1    1    1    1    1    1    1     1
#> [10,]    1    1    1    1    1    1    1    1    1     1
#> 
#> 
#> $sconc
#> $sconc[[1]]
#> RMODFLOW 3d array with 10 rows, 10 columns and 3 layers, representing the i, j & k dimensions. 
#> Not representing stress period data 
#> , , 1
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    0    0    0    0    0    0    0    0    0     0
#>  [2,]    0    0    0    0    0    0    0    0    0     0
#>  [3,]    0    0    0    0    0    0    0    0    0     0
#>  [4,]    0    0    0    0    0    0    0    0    0     0
#>  [5,]    0    0    0    0    0    0    0    0    0     0
#>  [6,]    0    0    0    0    0    0    0    0    0     0
#>  [7,]    0    0    0    0    0    0    0    0    0     0
#>  [8,]    0    0    0    0    0    0    0    0    0     0
#>  [9,]    0    0    0    0    0    0    0    0    0     0
#> [10,]    0    0    0    0    0    0    0    0    0     0
#> 
#> , , 2
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    0    0    0    0    0    0    0    0    0     0
#>  [2,]    0    0    0    0    0    0    0    0    0     0
#>  [3,]    0    0    0    0    0    0    0    0    0     0
#>  [4,]    0    0    0    0    0    0    0    0    0     0
#>  [5,]    0    0    0    0    0    0    0    0    0     0
#>  [6,]    0    0    0    0    0    0    0    0    0     0
#>  [7,]    0    0    0    0    0    0    0    0    0     0
#>  [8,]    0    0    0    0    0    0    0    0    0     0
#>  [9,]    0    0    0    0    0    0    0    0    0     0
#> [10,]    0    0    0    0    0    0    0    0    0     0
#> 
#> , , 3
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    0    0    0    0    0    0    0    0    0     0
#>  [2,]    0    0    0    0    0    0    0    0    0     0
#>  [3,]    0    0    0    0    0    0    0    0    0     0
#>  [4,]    0    0    0    0    0    0    0    0    0     0
#>  [5,]    0    0    0    0    0    0    0    0    0     0
#>  [6,]    0    0    0    0    0    0    0    0    0     0
#>  [7,]    0    0    0    0    0    0    0    0    0     0
#>  [8,]    0    0    0    0    0    0    0    0    0     0
#>  [9,]    0    0    0    0    0    0    0    0    0     0
#> [10,]    0    0    0    0    0    0    0    0    0     0
#> 
#> 
#> 
#> $cinact
#> [1] -888
#> 
#> $thkmin
#> [1] 0.01
#> 
#> $ifmtcn
#> [1] 0
#> 
#> $ifmtnp
#> [1] 0
#> 
#> $ifmtrf
#> [1] 0
#> 
#> $ifmtdp
#> [1] 0
#> 
#> $savucn
#> [1] TRUE
#> 
#> $nprs
#> [1] -1
#> 
#> $chkmas
#> [1] TRUE
#> 
#> $nprmas
#> [1] 1
#> 
#> $perlen
#> [1] 1
#> 
#> $nstp
#> [1] 1
#> 
#> $tsmult
#> [1] 1
#> 
#> $ssflag
#> [1] FALSE
#> 
#> $dt0
#> [1] 0
#> 
#> $mxstrn
#> [1] 50
#> 
#> $ttsmult
#> [1] 1
#> 
#> $ttsmax
#> [1] 0
#> 
#> attr(,"class")
#> [1] "btn"         "rmt_package"
#> 
#> $adv
#> $mixelm
#> [1] -1
#> 
#> $percel
#> [1] 1
#> 
#> $mxpart
#> [1] 1e+06
#> 
#> $nadvfd
#> [1] 0
#> 
#> attr(,"class")
#> [1] "adv"         "rmt_package"
#> 
#> $gcg
#> $mxiter
#> [1] 1
#> 
#> $iter1
#> [1] 50
#> 
#> $isolve
#> [1] 3
#> 
#> $ncrs
#> [1] 0
#> 
#> $accl
#> [1] 1
#> 
#> $cclose
#> [1] 1e-06
#> 
#> $iprgcg
#> [1] 0
#> 
#> attr(,"class")
#> [1] "gcg"         "rmt_package"
#> 
#> $nam
#>   ftype nunit      fname options
#> 1  LIST    50 output.m3d    <NA>
#> 2   FTL    51 output.ftl    FREE
#> 3   BTN    52  input.btn    <NA>
#> 4   ADV    53  input.adv    <NA>
#> 5   GCG    54  input.gcg    <NA>
#> 
#> attr(,"class")
#> [1] "mt3dms"

rmt_create(btn, adv, gcg, nam, recreate_nam = TRUE, basename = 'ex1')

#> $btn
#> $modflowstylearrays
#> [1] FALSE
#> 
#> $drycell
#> [1] FALSE
#> 
#> $legacy99storage
#> [1] FALSE
#> 
#> $ftlprint
#> [1] FALSE
#> 
#> $nowetdryprint
#> [1] FALSE
#> 
#> $omitdrycellbudget
#> [1] FALSE
#> 
#> $altwtsorb
#> [1] FALSE
#> 
#> $nlay
#> [1] 3
#> 
#> $nrow
#> [1] 10
#> 
#> $ncol
#> [1] 10
#> 
#> $nper
#> [1] 1
#> 
#> $ncomp
#> [1] 1
#> 
#> $mcomp
#> [1] 1
#> 
#> $tunit
#> [1] "s"
#> 
#> $lunit
#> [1] "m"
#> 
#> $munit
#> [1] "g"
#> 
#> $trnop
#>  [1]  TRUE  TRUE  TRUE  TRUE  TRUE FALSE FALSE FALSE FALSE FALSE
#> 
#> $laycon
#> [1] 1 1 1
#> 
#> $delr
#>  [1] 100 100 100 100 100 100 100 100 100 100
#> 
#> $delc
#>  [1] 100 100 100 100 100 100 100 100 100 100
#> 
#> $htop
#> RMODFLOW 2d array with 10 rows and 10 columns, representing the i & j dimensions. 
#> Not representing stress period data 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    0    0    0    0    0    0    0    0    0     0
#>  [2,]    0    0    0    0    0    0    0    0    0     0
#>  [3,]    0    0    0    0    0    0    0    0    0     0
#>  [4,]    0    0    0    0    0    0    0    0    0     0
#>  [5,]    0    0    0    0    0    0    0    0    0     0
#>  [6,]    0    0    0    0    0    0    0    0    0     0
#>  [7,]    0    0    0    0    0    0    0    0    0     0
#>  [8,]    0    0    0    0    0    0    0    0    0     0
#>  [9,]    0    0    0    0    0    0    0    0    0     0
#> [10,]    0    0    0    0    0    0    0    0    0     0
#> 
#> $dz
#> RMODFLOW 3d array with 10 rows, 10 columns and 3 layers, representing the i, j & k dimensions. 
#> Not representing stress period data 
#> , , 1
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]   10   10   10   10   10   10   10   10   10    10
#>  [2,]   10   10   10   10   10   10   10   10   10    10
#>  [3,]   10   10   10   10   10   10   10   10   10    10
#>  [4,]   10   10   10   10   10   10   10   10   10    10
#>  [5,]   10   10   10   10   10   10   10   10   10    10
#>  [6,]   10   10   10   10   10   10   10   10   10    10
#>  [7,]   10   10   10   10   10   10   10   10   10    10
#>  [8,]   10   10   10   10   10   10   10   10   10    10
#>  [9,]   10   10   10   10   10   10   10   10   10    10
#> [10,]   10   10   10   10   10   10   10   10   10    10
#> 
#> , , 2
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]   10   10   10   10   10   10   10   10   10    10
#>  [2,]   10   10   10   10   10   10   10   10   10    10
#>  [3,]   10   10   10   10   10   10   10   10   10    10
#>  [4,]   10   10   10   10   10   10   10   10   10    10
#>  [5,]   10   10   10   10   10   10   10   10   10    10
#>  [6,]   10   10   10   10   10   10   10   10   10    10
#>  [7,]   10   10   10   10   10   10   10   10   10    10
#>  [8,]   10   10   10   10   10   10   10   10   10    10
#>  [9,]   10   10   10   10   10   10   10   10   10    10
#> [10,]   10   10   10   10   10   10   10   10   10    10
#> 
#> , , 3
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]   10   10   10   10   10   10   10   10   10    10
#>  [2,]   10   10   10   10   10   10   10   10   10    10
#>  [3,]   10   10   10   10   10   10   10   10   10    10
#>  [4,]   10   10   10   10   10   10   10   10   10    10
#>  [5,]   10   10   10   10   10   10   10   10   10    10
#>  [6,]   10   10   10   10   10   10   10   10   10    10
#>  [7,]   10   10   10   10   10   10   10   10   10    10
#>  [8,]   10   10   10   10   10   10   10   10   10    10
#>  [9,]   10   10   10   10   10   10   10   10   10    10
#> [10,]   10   10   10   10   10   10   10   10   10    10
#> 
#> 
#> $prsity
#> RMODFLOW 3d array with 10 rows, 10 columns and 3 layers, representing the i, j & k dimensions. 
#> Not representing stress period data 
#> , , 1
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [2,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [3,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [4,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [5,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [6,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [7,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [8,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [9,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#> [10,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#> 
#> , , 2
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [2,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [3,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [4,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [5,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [6,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [7,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [8,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [9,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#> [10,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#> 
#> , , 3
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [2,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [3,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [4,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [5,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [6,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [7,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [8,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [9,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#> [10,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#> 
#> 
#> $icbund
#> RMODFLOW 3d array with 10 rows, 10 columns and 3 layers, representing the i, j & k dimensions. 
#> Not representing stress period data 
#> , , 1
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    1    1    1    1    1    1    1    1    1     1
#>  [2,]    1    1    1    1    1    1    1    1    1     1
#>  [3,]    1    1    1    1    1    1    1    1    1     1
#>  [4,]    1    1    1    1    1    1    1    1    1     1
#>  [5,]    1    1    1    1    1    1    1    1    1     1
#>  [6,]    1    1    1    1    1    1    1    1    1     1
#>  [7,]    1    1    1    1    1    1    1    1    1     1
#>  [8,]    1    1    1    1    1    1    1    1    1     1
#>  [9,]    1    1    1    1    1    1    1    1    1     1
#> [10,]    1    1    1    1    1    1    1    1    1     1
#> 
#> , , 2
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    1    1    1    1    1    1    1    1    1     1
#>  [2,]    1    1    1    1    1    1    1    1    1     1
#>  [3,]    1    1    1    1    1    1    1    1    1     1
#>  [4,]    1    1    1    1    1    1    1    1    1     1
#>  [5,]    1    1    1    1    1    1    1    1    1     1
#>  [6,]    1    1    1    1    1    1    1    1    1     1
#>  [7,]    1    1    1    1    1    1    1    1    1     1
#>  [8,]    1    1    1    1    1    1    1    1    1     1
#>  [9,]    1    1    1    1    1    1    1    1    1     1
#> [10,]    1    1    1    1    1    1    1    1    1     1
#> 
#> , , 3
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    1    1    1    1    1    1    1    1    1     1
#>  [2,]    1    1    1    1    1    1    1    1    1     1
#>  [3,]    1    1    1    1    1    1    1    1    1     1
#>  [4,]    1    1    1    1    1    1    1    1    1     1
#>  [5,]    1    1    1    1    1    1    1    1    1     1
#>  [6,]    1    1    1    1    1    1    1    1    1     1
#>  [7,]    1    1    1    1    1    1    1    1    1     1
#>  [8,]    1    1    1    1    1    1    1    1    1     1
#>  [9,]    1    1    1    1    1    1    1    1    1     1
#> [10,]    1    1    1    1    1    1    1    1    1     1
#> 
#> 
#> $sconc
#> $sconc[[1]]
#> RMODFLOW 3d array with 10 rows, 10 columns and 3 layers, representing the i, j & k dimensions. 
#> Not representing stress period data 
#> , , 1
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    0    0    0    0    0    0    0    0    0     0
#>  [2,]    0    0    0    0    0    0    0    0    0     0
#>  [3,]    0    0    0    0    0    0    0    0    0     0
#>  [4,]    0    0    0    0    0    0    0    0    0     0
#>  [5,]    0    0    0    0    0    0    0    0    0     0
#>  [6,]    0    0    0    0    0    0    0    0    0     0
#>  [7,]    0    0    0    0    0    0    0    0    0     0
#>  [8,]    0    0    0    0    0    0    0    0    0     0
#>  [9,]    0    0    0    0    0    0    0    0    0     0
#> [10,]    0    0    0    0    0    0    0    0    0     0
#> 
#> , , 2
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    0    0    0    0    0    0    0    0    0     0
#>  [2,]    0    0    0    0    0    0    0    0    0     0
#>  [3,]    0    0    0    0    0    0    0    0    0     0
#>  [4,]    0    0    0    0    0    0    0    0    0     0
#>  [5,]    0    0    0    0    0    0    0    0    0     0
#>  [6,]    0    0    0    0    0    0    0    0    0     0
#>  [7,]    0    0    0    0    0    0    0    0    0     0
#>  [8,]    0    0    0    0    0    0    0    0    0     0
#>  [9,]    0    0    0    0    0    0    0    0    0     0
#> [10,]    0    0    0    0    0    0    0    0    0     0
#> 
#> , , 3
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    0    0    0    0    0    0    0    0    0     0
#>  [2,]    0    0    0    0    0    0    0    0    0     0
#>  [3,]    0    0    0    0    0    0    0    0    0     0
#>  [4,]    0    0    0    0    0    0    0    0    0     0
#>  [5,]    0    0    0    0    0    0    0    0    0     0
#>  [6,]    0    0    0    0    0    0    0    0    0     0
#>  [7,]    0    0    0    0    0    0    0    0    0     0
#>  [8,]    0    0    0    0    0    0    0    0    0     0
#>  [9,]    0    0    0    0    0    0    0    0    0     0
#> [10,]    0    0    0    0    0    0    0    0    0     0
#> 
#> 
#> 
#> $cinact
#> [1] -888
#> 
#> $thkmin
#> [1] 0.01
#> 
#> $ifmtcn
#> [1] 0
#> 
#> $ifmtnp
#> [1] 0
#> 
#> $ifmtrf
#> [1] 0
#> 
#> $ifmtdp
#> [1] 0
#> 
#> $savucn
#> [1] TRUE
#> 
#> $nprs
#> [1] -1
#> 
#> $chkmas
#> [1] TRUE
#> 
#> $nprmas
#> [1] 1
#> 
#> $perlen
#> [1] 1
#> 
#> $nstp
#> [1] 1
#> 
#> $tsmult
#> [1] 1
#> 
#> $ssflag
#> [1] FALSE
#> 
#> $dt0
#> [1] 0
#> 
#> $mxstrn
#> [1] 50
#> 
#> $ttsmult
#> [1] 1
#> 
#> $ttsmax
#> [1] 0
#> 
#> attr(,"class")
#> [1] "btn"         "rmt_package"
#> 
#> $adv
#> $mixelm
#> [1] -1
#> 
#> $percel
#> [1] 1
#> 
#> $mxpart
#> [1] 1e+06
#> 
#> $nadvfd
#> [1] 0
#> 
#> attr(,"class")
#> [1] "adv"         "rmt_package"
#> 
#> $gcg
#> $mxiter
#> [1] 1
#> 
#> $iter1
#> [1] 50
#> 
#> $isolve
#> [1] 3
#> 
#> $ncrs
#> [1] 0
#> 
#> $accl
#> [1] 1
#> 
#> $cclose
#> [1] 1e-06
#> 
#> $iprgcg
#> [1] 0
#> 
#> attr(,"class")
#> [1] "gcg"         "rmt_package"
#> 
#> $nam
#>   ftype nunit      fname options
#> 1  LIST    50    ex1.m3d    <NA>
#> 2   FTL    51 output.ftl    FREE
#> 3   BTN    52    ex1.btn    <NA>
#> 4   ADV    53    ex1.adv    <NA>
#> 5   GCG    54    ex1.gcg    <NA>
#> 
#> attr(,"class")
#> [1] "mt3dms"

rmt_create(btn, adv, gcg, ftl = 'output.ftl', ftl_free = FALSE)

#> $btn
#> $modflowstylearrays
#> [1] FALSE
#> 
#> $drycell
#> [1] FALSE
#> 
#> $legacy99storage
#> [1] FALSE
#> 
#> $ftlprint
#> [1] FALSE
#> 
#> $nowetdryprint
#> [1] FALSE
#> 
#> $omitdrycellbudget
#> [1] FALSE
#> 
#> $altwtsorb
#> [1] FALSE
#> 
#> $nlay
#> [1] 3
#> 
#> $nrow
#> [1] 10
#> 
#> $ncol
#> [1] 10
#> 
#> $nper
#> [1] 1
#> 
#> $ncomp
#> [1] 1
#> 
#> $mcomp
#> [1] 1
#> 
#> $tunit
#> [1] "s"
#> 
#> $lunit
#> [1] "m"
#> 
#> $munit
#> [1] "g"
#> 
#> $trnop
#>  [1]  TRUE  TRUE  TRUE  TRUE  TRUE FALSE FALSE FALSE FALSE FALSE
#> 
#> $laycon
#> [1] 1 1 1
#> 
#> $delr
#>  [1] 100 100 100 100 100 100 100 100 100 100
#> 
#> $delc
#>  [1] 100 100 100 100 100 100 100 100 100 100
#> 
#> $htop
#> RMODFLOW 2d array with 10 rows and 10 columns, representing the i & j dimensions. 
#> Not representing stress period data 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    0    0    0    0    0    0    0    0    0     0
#>  [2,]    0    0    0    0    0    0    0    0    0     0
#>  [3,]    0    0    0    0    0    0    0    0    0     0
#>  [4,]    0    0    0    0    0    0    0    0    0     0
#>  [5,]    0    0    0    0    0    0    0    0    0     0
#>  [6,]    0    0    0    0    0    0    0    0    0     0
#>  [7,]    0    0    0    0    0    0    0    0    0     0
#>  [8,]    0    0    0    0    0    0    0    0    0     0
#>  [9,]    0    0    0    0    0    0    0    0    0     0
#> [10,]    0    0    0    0    0    0    0    0    0     0
#> 
#> $dz
#> RMODFLOW 3d array with 10 rows, 10 columns and 3 layers, representing the i, j & k dimensions. 
#> Not representing stress period data 
#> , , 1
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]   10   10   10   10   10   10   10   10   10    10
#>  [2,]   10   10   10   10   10   10   10   10   10    10
#>  [3,]   10   10   10   10   10   10   10   10   10    10
#>  [4,]   10   10   10   10   10   10   10   10   10    10
#>  [5,]   10   10   10   10   10   10   10   10   10    10
#>  [6,]   10   10   10   10   10   10   10   10   10    10
#>  [7,]   10   10   10   10   10   10   10   10   10    10
#>  [8,]   10   10   10   10   10   10   10   10   10    10
#>  [9,]   10   10   10   10   10   10   10   10   10    10
#> [10,]   10   10   10   10   10   10   10   10   10    10
#> 
#> , , 2
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]   10   10   10   10   10   10   10   10   10    10
#>  [2,]   10   10   10   10   10   10   10   10   10    10
#>  [3,]   10   10   10   10   10   10   10   10   10    10
#>  [4,]   10   10   10   10   10   10   10   10   10    10
#>  [5,]   10   10   10   10   10   10   10   10   10    10
#>  [6,]   10   10   10   10   10   10   10   10   10    10
#>  [7,]   10   10   10   10   10   10   10   10   10    10
#>  [8,]   10   10   10   10   10   10   10   10   10    10
#>  [9,]   10   10   10   10   10   10   10   10   10    10
#> [10,]   10   10   10   10   10   10   10   10   10    10
#> 
#> , , 3
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]   10   10   10   10   10   10   10   10   10    10
#>  [2,]   10   10   10   10   10   10   10   10   10    10
#>  [3,]   10   10   10   10   10   10   10   10   10    10
#>  [4,]   10   10   10   10   10   10   10   10   10    10
#>  [5,]   10   10   10   10   10   10   10   10   10    10
#>  [6,]   10   10   10   10   10   10   10   10   10    10
#>  [7,]   10   10   10   10   10   10   10   10   10    10
#>  [8,]   10   10   10   10   10   10   10   10   10    10
#>  [9,]   10   10   10   10   10   10   10   10   10    10
#> [10,]   10   10   10   10   10   10   10   10   10    10
#> 
#> 
#> $prsity
#> RMODFLOW 3d array with 10 rows, 10 columns and 3 layers, representing the i, j & k dimensions. 
#> Not representing stress period data 
#> , , 1
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [2,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [3,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [4,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [5,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [6,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [7,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [8,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [9,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#> [10,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#> 
#> , , 2
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [2,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [3,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [4,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [5,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [6,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [7,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [8,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [9,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#> [10,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#> 
#> , , 3
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [2,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [3,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [4,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [5,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [6,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [7,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [8,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [9,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#> [10,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#> 
#> 
#> $icbund
#> RMODFLOW 3d array with 10 rows, 10 columns and 3 layers, representing the i, j & k dimensions. 
#> Not representing stress period data 
#> , , 1
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    1    1    1    1    1    1    1    1    1     1
#>  [2,]    1    1    1    1    1    1    1    1    1     1
#>  [3,]    1    1    1    1    1    1    1    1    1     1
#>  [4,]    1    1    1    1    1    1    1    1    1     1
#>  [5,]    1    1    1    1    1    1    1    1    1     1
#>  [6,]    1    1    1    1    1    1    1    1    1     1
#>  [7,]    1    1    1    1    1    1    1    1    1     1
#>  [8,]    1    1    1    1    1    1    1    1    1     1
#>  [9,]    1    1    1    1    1    1    1    1    1     1
#> [10,]    1    1    1    1    1    1    1    1    1     1
#> 
#> , , 2
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    1    1    1    1    1    1    1    1    1     1
#>  [2,]    1    1    1    1    1    1    1    1    1     1
#>  [3,]    1    1    1    1    1    1    1    1    1     1
#>  [4,]    1    1    1    1    1    1    1    1    1     1
#>  [5,]    1    1    1    1    1    1    1    1    1     1
#>  [6,]    1    1    1    1    1    1    1    1    1     1
#>  [7,]    1    1    1    1    1    1    1    1    1     1
#>  [8,]    1    1    1    1    1    1    1    1    1     1
#>  [9,]    1    1    1    1    1    1    1    1    1     1
#> [10,]    1    1    1    1    1    1    1    1    1     1
#> 
#> , , 3
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    1    1    1    1    1    1    1    1    1     1
#>  [2,]    1    1    1    1    1    1    1    1    1     1
#>  [3,]    1    1    1    1    1    1    1    1    1     1
#>  [4,]    1    1    1    1    1    1    1    1    1     1
#>  [5,]    1    1    1    1    1    1    1    1    1     1
#>  [6,]    1    1    1    1    1    1    1    1    1     1
#>  [7,]    1    1    1    1    1    1    1    1    1     1
#>  [8,]    1    1    1    1    1    1    1    1    1     1
#>  [9,]    1    1    1    1    1    1    1    1    1     1
#> [10,]    1    1    1    1    1    1    1    1    1     1
#> 
#> 
#> $sconc
#> $sconc[[1]]
#> RMODFLOW 3d array with 10 rows, 10 columns and 3 layers, representing the i, j & k dimensions. 
#> Not representing stress period data 
#> , , 1
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    0    0    0    0    0    0    0    0    0     0
#>  [2,]    0    0    0    0    0    0    0    0    0     0
#>  [3,]    0    0    0    0    0    0    0    0    0     0
#>  [4,]    0    0    0    0    0    0    0    0    0     0
#>  [5,]    0    0    0    0    0    0    0    0    0     0
#>  [6,]    0    0    0    0    0    0    0    0    0     0
#>  [7,]    0    0    0    0    0    0    0    0    0     0
#>  [8,]    0    0    0    0    0    0    0    0    0     0
#>  [9,]    0    0    0    0    0    0    0    0    0     0
#> [10,]    0    0    0    0    0    0    0    0    0     0
#> 
#> , , 2
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    0    0    0    0    0    0    0    0    0     0
#>  [2,]    0    0    0    0    0    0    0    0    0     0
#>  [3,]    0    0    0    0    0    0    0    0    0     0
#>  [4,]    0    0    0    0    0    0    0    0    0     0
#>  [5,]    0    0    0    0    0    0    0    0    0     0
#>  [6,]    0    0    0    0    0    0    0    0    0     0
#>  [7,]    0    0    0    0    0    0    0    0    0     0
#>  [8,]    0    0    0    0    0    0    0    0    0     0
#>  [9,]    0    0    0    0    0    0    0    0    0     0
#> [10,]    0    0    0    0    0    0    0    0    0     0
#> 
#> , , 3
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    0    0    0    0    0    0    0    0    0     0
#>  [2,]    0    0    0    0    0    0    0    0    0     0
#>  [3,]    0    0    0    0    0    0    0    0    0     0
#>  [4,]    0    0    0    0    0    0    0    0    0     0
#>  [5,]    0    0    0    0    0    0    0    0    0     0
#>  [6,]    0    0    0    0    0    0    0    0    0     0
#>  [7,]    0    0    0    0    0    0    0    0    0     0
#>  [8,]    0    0    0    0    0    0    0    0    0     0
#>  [9,]    0    0    0    0    0    0    0    0    0     0
#> [10,]    0    0    0    0    0    0    0    0    0     0
#> 
#> 
#> 
#> $cinact
#> [1] -888
#> 
#> $thkmin
#> [1] 0.01
#> 
#> $ifmtcn
#> [1] 0
#> 
#> $ifmtnp
#> [1] 0
#> 
#> $ifmtrf
#> [1] 0
#> 
#> $ifmtdp
#> [1] 0
#> 
#> $savucn
#> [1] TRUE
#> 
#> $nprs
#> [1] -1
#> 
#> $chkmas
#> [1] TRUE
#> 
#> $nprmas
#> [1] 1
#> 
#> $perlen
#> [1] 1
#> 
#> $nstp
#> [1] 1
#> 
#> $tsmult
#> [1] 1
#> 
#> $ssflag
#> [1] FALSE
#> 
#> $dt0
#> [1] 0
#> 
#> $mxstrn
#> [1] 50
#> 
#> $ttsmult
#> [1] 1
#> 
#> $ttsmax
#> [1] 0
#> 
#> attr(,"class")
#> [1] "btn"         "rmt_package"
#> 
#> $adv
#> $mixelm
#> [1] -1
#> 
#> $percel
#> [1] 1
#> 
#> $mxpart
#> [1] 1e+06
#> 
#> $nadvfd
#> [1] 0
#> 
#> attr(,"class")
#> [1] "adv"         "rmt_package"
#> 
#> $gcg
#> $mxiter
#> [1] 1
#> 
#> $iter1
#> [1] 50
#> 
#> $isolve
#> [1] 3
#> 
#> $ncrs
#> [1] 0
#> 
#> $accl
#> [1] 1
#> 
#> $cclose
#> [1] 1e-06
#> 
#> $iprgcg
#> [1] 0
#> 
#> attr(,"class")
#> [1] "gcg"         "rmt_package"
#> 
#> $nam
#>   ftype nunit      fname options
#> 1  LIST    50 output.m3d      NA
#> 2   FTL    51 output.ftl      NA
#> 3   BTN    52  input.btn      NA
#> 4   ADV    53  input.adv      NA
#> 5   GCG    54  input.gcg      NA
#> 
#> attr(,"class")
#> [1] "mt3dms"