Create an RMT3DMS adv object

rmt_create_adv creates an RMT3DMS adv object.

rmt_create_adv(
  mixelm = -1,
  percel = 1,
  mxpart = 1e+06,
  nadvfd = 0,
  itrack = 1,
  wd = 0.5,
  dceps = 1e-05,
  nplane = 0,
  npl = 0,
  nph = 32,
  npmin = 2,
  npmax = 2 * nph,
  interp = 1,
  nlsink = nplane,
  npsink = nph,
  dchmoc = 0.001
)

Arguments

mixelm
integer code for the advection solution option (see details). Defaults to -1.

percel
Courant number. Defaults to 1

mxpart
maximum number of moving particles. Only used when mixelm = 1 or 3. Defaults to 1e6

nadvfd
weighting scheme for implicit finite difference (mixelm = 0). 0 or 1 (default) indicates upstream weighting, 2 is central-in-space

itrack
particle-tracking method: 1 (default) = first-order Euler, 2 = fourth-order Runge-Kutta, 3 = hybrid. Only used when mixelm = 1, 2 or 3.

wd
concentration weighting factor used for operator splitting in particle tracking-based methods. Defaults to 0.5. Only used when mixelm = 1, 2 or 3.

dceps
Relative Cell Concentration Gradient below which advective transport is negligble. Defaults to 1e-5. Only used when mixelm = 1 or 3.

nplane
specifies type of initial placement of particles: random (0; default) or fixed pattern (> 0). Only used when mixelm = 1 or 3.

npl
number of initial particles per cell at cells where Relative Cell Concentration Gradient is <= dceps. Defaults to 0. Only used when mixelm = 1 or 3.

nph
number of initial particles per cell at cells where Relative Cell Concentration Gradient is > dceps. Defaults to 32. Only used when mixelm = 1 or 3.

npmin
minimum number of particles allowed per cell. Defaults to 2. Only used when mixelm = 1 or 3.

npmax
maximum number of particles allowed per cell. Defaults to 2 * nph. Only used when mixelm = 1 or 3.

interp
concentration interpolation method for MMOC scheme. Only a value of 1 is allowed (linear interpolation). Only used when mixelm = 2 or 3.

nlsink
specifies type of initial placement of particles to approximate sink cells in MMOC: random (0) or fixed pattern (> 0). Defaults to nplane. Only used when mixelm = 2 or 3.

npsink
number of particles used to approximate sink cells in MMOC. Defaults to nph. Only used when mixelm = 2 or 3.

dchmoc
critical Relative Concentration Gradient for controlling selective use of MOC or MMOC in HMOC scheme. Defaults to 0.001. Only used when mixelm = 3.

Value

an object of class adv

Details

mixelm = 0: standard finite-difference method with upstream (nadvfd = 0 or 1) or central-in-space (nadvfd = 2) weighting. mixelm = 1: forward-tracking method of characteristics (MOC) mixelm = 2: backward-tracking modified method of characteristics (MMOC) mixelm = 3: hybrid method of characteristics (HMOC) with MOC or MMOC automatically and dynamically selected mixelm = -1: third-order TVD scheme (ULTIMATE)

See also

Examples

rmt_create_adv(percel = 0.75)

#> $mixelm
#> [1] -1
#> 
#> $percel
#> [1] 0.75
#> 
#> $mxpart
#> [1] 1e+06
#> 
#> $nadvfd
#> [1] 0
#> 
#> attr(,"class")
#> [1] "adv"         "rmt_package"

rmt_create_adv(mixelm = 3, nph = 16)

#> $mixelm
#> [1] 3
#> 
#> $percel
#> [1] 1
#> 
#> $mxpart
#> [1] 1e+06
#> 
#> $nadvfd
#> [1] 0
#> 
#> $itrack
#> [1] 1
#> 
#> $wd
#> [1] 0.5
#> 
#> $dceps
#> [1] 1e-05
#> 
#> $nplane
#> [1] 0
#> 
#> $npl
#> [1] 0
#> 
#> $nph
#> [1] 16
#> 
#> $npmin
#> [1] 2
#> 
#> $npmax
#> [1] 32
#> 
#> $interp
#> [1] 1
#> 
#> $nlsink
#> [1] 0
#> 
#> $npsink
#> [1] 16
#> 
#> $dchmoc
#> [1] 0.001
#> 
#> attr(,"class")
#> [1] "adv"         "rmt_package"