Create a RMT3DMS btn object

rmt_create_btn creates a RMT3DMS btn object.

rmt_create_btn(
  nlay = 3,
  nrow = 10,
  ncol = 10,
  nper = 1,
  ncomp = 1,
  mcomp = ncomp,
  tunit = "s",
  lunit = "m",
  munit = "g",
  trnop = c(rep(TRUE, 5), rep(FALSE, 5)),
  laycon = rep(1, nlay),
  delr = 100,
  delc = 100,
  htop = 0,
  dz = 10,
  prsity = 0.3,
  icbund = 1,
  sconc = as.list(rep(0, ncomp)),
  cinact = -888,
  thkmin = 0.01,
  ifmtcn = 0,
  ifmtnp = 0,
  ifmtrf = 0,
  ifmtdp = 0,
  savucn = TRUE,
  nprs = -1,
  timprs = NULL,
  obs = NULL,
  nprobs = 1,
  chkmas = TRUE,
  nprmas = 1,
  perlen = rep(1, nper),
  nstp = rep(1, nper),
  tsmult = rep(1, nper),
  ssflag = rep(FALSE, nper),
  tslngh = lapply(seq_len(nper), function(i) rep(1, nstp[i])),
  dt0 = rep(0, nper),
  mxstrn = rep(50, nper),
  ttsmult = rep(1, nper),
  ttsmax = rep(0, nper),
  modflowstylearrays = FALSE,
  drycell = FALSE,
  legacy99storage = FALSE,
  ftlprint = FALSE,
  nowetdryprint = FALSE,
  omitdrycellbudget = FALSE,
  altwtsorb = FALSE,
  prj = NULL
)

Arguments

nlay
number of layers; defaults to 3

nrow
number of rows; defaults to 10

ncol
number of columns; defaults to 10

nper
number of stress-periods in the flow model; defaults to 1

ncomp
total number of chemical species in the simulation; defaults to 1

mcomp
total number of mobile species; defaults to ncomp

tunit
character; time unit. Only for identification purposes. Defaults to 's'.

lunit
character; length unit. Only for identification purposes. Defaults to 'm'.

munit
character: mass unit. Only for identification purposes. Defaults to 'mg'.

trnop
logicals denoting which packages are active. Not used by MT3DMS > 4.50 or by MT3D-USGS. Defaults to setting adv, dsp, ssm, rct & gcg active.

laycon
vector indicating type of model layer: confined (0) or convertible (not 0; default for all layers).

delr
vector of cell widths along rows; defaults to 100 for all columns

delc
vector of cell widths along columns; defaults to 100 for all rows

htop
matrix with the top elevation of layer 1; defaults to 0 for all cells

dz
3D array with the thicknesses of each cell. Defaults to 10 for all cells.

prsity
3D array with the (effective or mobile) porosity values for each cell. Defaults to 0.3 for all cells.

icbund
3D array with integer values for each cell specifying if solute transport is active (1), inactive (0) or if the cell is a constant-concentration (-1). Defaults to 1 for all cells.

sconc
list with ncomp 3D arrays specifying the starting concentrations in each cell for each species. Defaults to 0 for all cells for all species.

cinact
value for indicating an inactive concentration cell (icbund = 0). Defaults to -888.

thkmin
decimal fraction of layer thickness below which the cell is considered inactive. Defaults to 0.01. If negative, the absolute value is used.

ifmtcn
flag and printing code to determine if calculated concentration should be printed to the standard output text file. Defaults to 0 (no printing).

ifmtnp
flag and printing code to determine if number of particles in each cell should be printed to the standard output text file. Defaults to 0 (no printing).

ifmtrf
flag and printing code to determine if the model-calculated retardation factor should be printed to the standard output text file. Defaults to 0 (no printing).

ifmtdp
flag and printing code to determine if the model-calculated distance-weighted dispersion coefficient should be printed to the standard output text file. Defaults to 0 (no printing).

savucn
logical determining if concentration solution should be saved in default unformatted (binary) file named MT3Dnnn.UCN where nnn is the species index number. Defaults to TRUE.

nprs
frequency of output. Ignored if timprs is supplied. If positive, timpr must be supplied. If negative, output is printed or saved whenever the number of transport steps is an even multiple of nprs. If zero, only print or save at the end of the simulation. Defaults to -1.

timprs
vector with user-defined printing/saving times. Defaults to NULL.

obs
optional data.frame with k, i and j columns indicating the cells for which species' concentrations will be saved in MT3Dnnn.OBS files (nnn indicating species index). Defaults to NULL.

nprobs
integer indicating how frequently concentrations at obs observation points should be saved. Concentrations are saved every nprobs step. Defaults to 1.

chkmas
logical indicating whether a one-line summary of the mass balance should be saved to a default MT3Dnnn.MAS file where nnn is the species index. Defaults to TRUE.

nprmas
frequency of saving the mass budget information to MT3Dnnn.MAS. Information is saved every nprmas step. Defaults to 1.

perlen
vector of flow model stress-period lengths. Defaults to 1 for every stress-period.

nstp
vector of flow model stress-period time steps. Defaults to 1 for every stress-period.

tsmult
vector of successive flow model time step length multipliers. Defaults to 1 for every stress-period.

ssflag
logical vector of length nper indicating if the steady-state transport option should be used. Defaults to FALSE for every stress-period.

tslngh
list of length nper with time step length vectors. Defaults to 1 for every time step length. Only used when tsmult > 0 for the current stress-period.

dt0
vector of length nper of initial transport step sizes within each flow time step. If zero (default) and the GCG solver is active, a model-calculated value based on the Courant number in the Advection package is used.

mxstrn
vector of length nper with the maximum number of transport steps allowed within one flow time step. Defaults to 50 for every stress-period.

ttsmult
vector of length nper with transport step size multipliers within a single flow time step. Defaults to 1 for every stress-period.

ttsmax
vector of length nperwith maximum allowed transport step size when ttsmult > 1. Defaults to 0 for every stress-period (no limit).

modflowstylearrays
logical; should MODFLOW-style arrays and array headers be used. Defaults to FALSE. MT3D-USGS only.

drycell
logical; should mass transfer through dry cells be enabled when dry cells remain active in the flow simulation (e.g. with MODFLOW-NWT). Only available when mixelm = 0 or -1 in the Advection package. Defaults to FALSE. MT3D-USGS only.

legacy99storage
logical; should MT3DMS flow storage computations be used. Defaults to FALSE. MT3D-USGS only.

ftlprint
logical; should the contents of the flow-transport link file be printed to the standard output text file. Defaults to FALSE. MT3D-USGS only.

nowetdryprint
logical; should printing of messages related to drying and re-wetting of model cells to the standard output text file be disabled? Defaults to FALSE. MT3D-USGS only.

omitdrycellbudget
logical; should the printing of the mass budget of dry cells be disabled? Defaults to FALSE. MT3D-USGS only.

altwtsorb
logical; use an alternative formulation to simulate adsorbed mass. Defaults to FALSE. MT3D-USGS only.

prj
RMODFLOW prj object. Defaults to NULL.

Value

an object of class btn

See also

Examples

rmt_create_btn()

#> $modflowstylearrays
#> [1] FALSE
#> 
#> $drycell
#> [1] FALSE
#> 
#> $legacy99storage
#> [1] FALSE
#> 
#> $ftlprint
#> [1] FALSE
#> 
#> $nowetdryprint
#> [1] FALSE
#> 
#> $omitdrycellbudget
#> [1] FALSE
#> 
#> $altwtsorb
#> [1] FALSE
#> 
#> $nlay
#> [1] 3
#> 
#> $nrow
#> [1] 10
#> 
#> $ncol
#> [1] 10
#> 
#> $nper
#> [1] 1
#> 
#> $ncomp
#> [1] 1
#> 
#> $mcomp
#> [1] 1
#> 
#> $tunit
#> [1] "s"
#> 
#> $lunit
#> [1] "m"
#> 
#> $munit
#> [1] "g"
#> 
#> $trnop
#>  [1]  TRUE  TRUE  TRUE  TRUE  TRUE FALSE FALSE FALSE FALSE FALSE
#> 
#> $laycon
#> [1] 1 1 1
#> 
#> $delr
#>  [1] 100 100 100 100 100 100 100 100 100 100
#> 
#> $delc
#>  [1] 100 100 100 100 100 100 100 100 100 100
#> 
#> $htop
#> RMODFLOW 2d array with 10 rows and 10 columns, representing the i & j dimensions. 
#> Not representing stress period data 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    0    0    0    0    0    0    0    0    0     0
#>  [2,]    0    0    0    0    0    0    0    0    0     0
#>  [3,]    0    0    0    0    0    0    0    0    0     0
#>  [4,]    0    0    0    0    0    0    0    0    0     0
#>  [5,]    0    0    0    0    0    0    0    0    0     0
#>  [6,]    0    0    0    0    0    0    0    0    0     0
#>  [7,]    0    0    0    0    0    0    0    0    0     0
#>  [8,]    0    0    0    0    0    0    0    0    0     0
#>  [9,]    0    0    0    0    0    0    0    0    0     0
#> [10,]    0    0    0    0    0    0    0    0    0     0
#> 
#> $dz
#> RMODFLOW 3d array with 10 rows, 10 columns and 3 layers, representing the i, j & k dimensions. 
#> Not representing stress period data 
#> , , 1
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]   10   10   10   10   10   10   10   10   10    10
#>  [2,]   10   10   10   10   10   10   10   10   10    10
#>  [3,]   10   10   10   10   10   10   10   10   10    10
#>  [4,]   10   10   10   10   10   10   10   10   10    10
#>  [5,]   10   10   10   10   10   10   10   10   10    10
#>  [6,]   10   10   10   10   10   10   10   10   10    10
#>  [7,]   10   10   10   10   10   10   10   10   10    10
#>  [8,]   10   10   10   10   10   10   10   10   10    10
#>  [9,]   10   10   10   10   10   10   10   10   10    10
#> [10,]   10   10   10   10   10   10   10   10   10    10
#> 
#> , , 2
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]   10   10   10   10   10   10   10   10   10    10
#>  [2,]   10   10   10   10   10   10   10   10   10    10
#>  [3,]   10   10   10   10   10   10   10   10   10    10
#>  [4,]   10   10   10   10   10   10   10   10   10    10
#>  [5,]   10   10   10   10   10   10   10   10   10    10
#>  [6,]   10   10   10   10   10   10   10   10   10    10
#>  [7,]   10   10   10   10   10   10   10   10   10    10
#>  [8,]   10   10   10   10   10   10   10   10   10    10
#>  [9,]   10   10   10   10   10   10   10   10   10    10
#> [10,]   10   10   10   10   10   10   10   10   10    10
#> 
#> , , 3
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]   10   10   10   10   10   10   10   10   10    10
#>  [2,]   10   10   10   10   10   10   10   10   10    10
#>  [3,]   10   10   10   10   10   10   10   10   10    10
#>  [4,]   10   10   10   10   10   10   10   10   10    10
#>  [5,]   10   10   10   10   10   10   10   10   10    10
#>  [6,]   10   10   10   10   10   10   10   10   10    10
#>  [7,]   10   10   10   10   10   10   10   10   10    10
#>  [8,]   10   10   10   10   10   10   10   10   10    10
#>  [9,]   10   10   10   10   10   10   10   10   10    10
#> [10,]   10   10   10   10   10   10   10   10   10    10
#> 
#> 
#> $prsity
#> RMODFLOW 3d array with 10 rows, 10 columns and 3 layers, representing the i, j & k dimensions. 
#> Not representing stress period data 
#> , , 1
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [2,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [3,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [4,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [5,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [6,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [7,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [8,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [9,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#> [10,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#> 
#> , , 2
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [2,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [3,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [4,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [5,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [6,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [7,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [8,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [9,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#> [10,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#> 
#> , , 3
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [2,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [3,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [4,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [5,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [6,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [7,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [8,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [9,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#> [10,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#> 
#> 
#> $icbund
#> RMODFLOW 3d array with 10 rows, 10 columns and 3 layers, representing the i, j & k dimensions. 
#> Not representing stress period data 
#> , , 1
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    1    1    1    1    1    1    1    1    1     1
#>  [2,]    1    1    1    1    1    1    1    1    1     1
#>  [3,]    1    1    1    1    1    1    1    1    1     1
#>  [4,]    1    1    1    1    1    1    1    1    1     1
#>  [5,]    1    1    1    1    1    1    1    1    1     1
#>  [6,]    1    1    1    1    1    1    1    1    1     1
#>  [7,]    1    1    1    1    1    1    1    1    1     1
#>  [8,]    1    1    1    1    1    1    1    1    1     1
#>  [9,]    1    1    1    1    1    1    1    1    1     1
#> [10,]    1    1    1    1    1    1    1    1    1     1
#> 
#> , , 2
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    1    1    1    1    1    1    1    1    1     1
#>  [2,]    1    1    1    1    1    1    1    1    1     1
#>  [3,]    1    1    1    1    1    1    1    1    1     1
#>  [4,]    1    1    1    1    1    1    1    1    1     1
#>  [5,]    1    1    1    1    1    1    1    1    1     1
#>  [6,]    1    1    1    1    1    1    1    1    1     1
#>  [7,]    1    1    1    1    1    1    1    1    1     1
#>  [8,]    1    1    1    1    1    1    1    1    1     1
#>  [9,]    1    1    1    1    1    1    1    1    1     1
#> [10,]    1    1    1    1    1    1    1    1    1     1
#> 
#> , , 3
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    1    1    1    1    1    1    1    1    1     1
#>  [2,]    1    1    1    1    1    1    1    1    1     1
#>  [3,]    1    1    1    1    1    1    1    1    1     1
#>  [4,]    1    1    1    1    1    1    1    1    1     1
#>  [5,]    1    1    1    1    1    1    1    1    1     1
#>  [6,]    1    1    1    1    1    1    1    1    1     1
#>  [7,]    1    1    1    1    1    1    1    1    1     1
#>  [8,]    1    1    1    1    1    1    1    1    1     1
#>  [9,]    1    1    1    1    1    1    1    1    1     1
#> [10,]    1    1    1    1    1    1    1    1    1     1
#> 
#> 
#> $sconc
#> $sconc[[1]]
#> RMODFLOW 3d array with 10 rows, 10 columns and 3 layers, representing the i, j & k dimensions. 
#> Not representing stress period data 
#> , , 1
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    0    0    0    0    0    0    0    0    0     0
#>  [2,]    0    0    0    0    0    0    0    0    0     0
#>  [3,]    0    0    0    0    0    0    0    0    0     0
#>  [4,]    0    0    0    0    0    0    0    0    0     0
#>  [5,]    0    0    0    0    0    0    0    0    0     0
#>  [6,]    0    0    0    0    0    0    0    0    0     0
#>  [7,]    0    0    0    0    0    0    0    0    0     0
#>  [8,]    0    0    0    0    0    0    0    0    0     0
#>  [9,]    0    0    0    0    0    0    0    0    0     0
#> [10,]    0    0    0    0    0    0    0    0    0     0
#> 
#> , , 2
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    0    0    0    0    0    0    0    0    0     0
#>  [2,]    0    0    0    0    0    0    0    0    0     0
#>  [3,]    0    0    0    0    0    0    0    0    0     0
#>  [4,]    0    0    0    0    0    0    0    0    0     0
#>  [5,]    0    0    0    0    0    0    0    0    0     0
#>  [6,]    0    0    0    0    0    0    0    0    0     0
#>  [7,]    0    0    0    0    0    0    0    0    0     0
#>  [8,]    0    0    0    0    0    0    0    0    0     0
#>  [9,]    0    0    0    0    0    0    0    0    0     0
#> [10,]    0    0    0    0    0    0    0    0    0     0
#> 
#> , , 3
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    0    0    0    0    0    0    0    0    0     0
#>  [2,]    0    0    0    0    0    0    0    0    0     0
#>  [3,]    0    0    0    0    0    0    0    0    0     0
#>  [4,]    0    0    0    0    0    0    0    0    0     0
#>  [5,]    0    0    0    0    0    0    0    0    0     0
#>  [6,]    0    0    0    0    0    0    0    0    0     0
#>  [7,]    0    0    0    0    0    0    0    0    0     0
#>  [8,]    0    0    0    0    0    0    0    0    0     0
#>  [9,]    0    0    0    0    0    0    0    0    0     0
#> [10,]    0    0    0    0    0    0    0    0    0     0
#> 
#> 
#> 
#> $cinact
#> [1] -888
#> 
#> $thkmin
#> [1] 0.01
#> 
#> $ifmtcn
#> [1] 0
#> 
#> $ifmtnp
#> [1] 0
#> 
#> $ifmtrf
#> [1] 0
#> 
#> $ifmtdp
#> [1] 0
#> 
#> $savucn
#> [1] TRUE
#> 
#> $nprs
#> [1] -1
#> 
#> $chkmas
#> [1] TRUE
#> 
#> $nprmas
#> [1] 1
#> 
#> $perlen
#> [1] 1
#> 
#> $nstp
#> [1] 1
#> 
#> $tsmult
#> [1] 1
#> 
#> $ssflag
#> [1] FALSE
#> 
#> $dt0
#> [1] 0
#> 
#> $mxstrn
#> [1] 50
#> 
#> $ttsmult
#> [1] 1
#> 
#> $ttsmax
#> [1] 0
#> 
#> attr(,"class")
#> [1] "btn"         "rmt_package"

prsity <- array(rep(c(0.1, 0.01, 0.3), each = 100), dim = c(10, 10, 3))
rmt_create_btn(nper = 3, prsity = prsity)

#> $modflowstylearrays
#> [1] FALSE
#> 
#> $drycell
#> [1] FALSE
#> 
#> $legacy99storage
#> [1] FALSE
#> 
#> $ftlprint
#> [1] FALSE
#> 
#> $nowetdryprint
#> [1] FALSE
#> 
#> $omitdrycellbudget
#> [1] FALSE
#> 
#> $altwtsorb
#> [1] FALSE
#> 
#> $nlay
#> [1] 3
#> 
#> $nrow
#> [1] 10
#> 
#> $ncol
#> [1] 10
#> 
#> $nper
#> [1] 3
#> 
#> $ncomp
#> [1] 1
#> 
#> $mcomp
#> [1] 1
#> 
#> $tunit
#> [1] "s"
#> 
#> $lunit
#> [1] "m"
#> 
#> $munit
#> [1] "g"
#> 
#> $trnop
#>  [1]  TRUE  TRUE  TRUE  TRUE  TRUE FALSE FALSE FALSE FALSE FALSE
#> 
#> $laycon
#> [1] 1 1 1
#> 
#> $delr
#>  [1] 100 100 100 100 100 100 100 100 100 100
#> 
#> $delc
#>  [1] 100 100 100 100 100 100 100 100 100 100
#> 
#> $htop
#> RMODFLOW 2d array with 10 rows and 10 columns, representing the i & j dimensions. 
#> Not representing stress period data 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    0    0    0    0    0    0    0    0    0     0
#>  [2,]    0    0    0    0    0    0    0    0    0     0
#>  [3,]    0    0    0    0    0    0    0    0    0     0
#>  [4,]    0    0    0    0    0    0    0    0    0     0
#>  [5,]    0    0    0    0    0    0    0    0    0     0
#>  [6,]    0    0    0    0    0    0    0    0    0     0
#>  [7,]    0    0    0    0    0    0    0    0    0     0
#>  [8,]    0    0    0    0    0    0    0    0    0     0
#>  [9,]    0    0    0    0    0    0    0    0    0     0
#> [10,]    0    0    0    0    0    0    0    0    0     0
#> 
#> $dz
#> RMODFLOW 3d array with 10 rows, 10 columns and 3 layers, representing the i, j & k dimensions. 
#> Not representing stress period data 
#> , , 1
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]   10   10   10   10   10   10   10   10   10    10
#>  [2,]   10   10   10   10   10   10   10   10   10    10
#>  [3,]   10   10   10   10   10   10   10   10   10    10
#>  [4,]   10   10   10   10   10   10   10   10   10    10
#>  [5,]   10   10   10   10   10   10   10   10   10    10
#>  [6,]   10   10   10   10   10   10   10   10   10    10
#>  [7,]   10   10   10   10   10   10   10   10   10    10
#>  [8,]   10   10   10   10   10   10   10   10   10    10
#>  [9,]   10   10   10   10   10   10   10   10   10    10
#> [10,]   10   10   10   10   10   10   10   10   10    10
#> 
#> , , 2
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]   10   10   10   10   10   10   10   10   10    10
#>  [2,]   10   10   10   10   10   10   10   10   10    10
#>  [3,]   10   10   10   10   10   10   10   10   10    10
#>  [4,]   10   10   10   10   10   10   10   10   10    10
#>  [5,]   10   10   10   10   10   10   10   10   10    10
#>  [6,]   10   10   10   10   10   10   10   10   10    10
#>  [7,]   10   10   10   10   10   10   10   10   10    10
#>  [8,]   10   10   10   10   10   10   10   10   10    10
#>  [9,]   10   10   10   10   10   10   10   10   10    10
#> [10,]   10   10   10   10   10   10   10   10   10    10
#> 
#> , , 3
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]   10   10   10   10   10   10   10   10   10    10
#>  [2,]   10   10   10   10   10   10   10   10   10    10
#>  [3,]   10   10   10   10   10   10   10   10   10    10
#>  [4,]   10   10   10   10   10   10   10   10   10    10
#>  [5,]   10   10   10   10   10   10   10   10   10    10
#>  [6,]   10   10   10   10   10   10   10   10   10    10
#>  [7,]   10   10   10   10   10   10   10   10   10    10
#>  [8,]   10   10   10   10   10   10   10   10   10    10
#>  [9,]   10   10   10   10   10   10   10   10   10    10
#> [10,]   10   10   10   10   10   10   10   10   10    10
#> 
#> 
#> $prsity
#> RMODFLOW 3d array with 10 rows, 10 columns and 3 layers, representing the i, j & k dimensions. 
#> Not representing stress period data 
#> , , 1
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1   0.1
#>  [2,]  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1   0.1
#>  [3,]  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1   0.1
#>  [4,]  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1   0.1
#>  [5,]  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1   0.1
#>  [6,]  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1   0.1
#>  [7,]  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1   0.1
#>  [8,]  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1   0.1
#>  [9,]  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1   0.1
#> [10,]  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1   0.1
#> 
#> , , 2
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01  0.01
#>  [2,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01  0.01
#>  [3,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01  0.01
#>  [4,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01  0.01
#>  [5,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01  0.01
#>  [6,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01  0.01
#>  [7,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01  0.01
#>  [8,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01  0.01
#>  [9,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01  0.01
#> [10,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01  0.01
#> 
#> , , 3
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [2,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [3,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [4,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [5,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [6,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [7,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [8,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [9,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#> [10,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#> 
#> 
#> $icbund
#> RMODFLOW 3d array with 10 rows, 10 columns and 3 layers, representing the i, j & k dimensions. 
#> Not representing stress period data 
#> , , 1
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    1    1    1    1    1    1    1    1    1     1
#>  [2,]    1    1    1    1    1    1    1    1    1     1
#>  [3,]    1    1    1    1    1    1    1    1    1     1
#>  [4,]    1    1    1    1    1    1    1    1    1     1
#>  [5,]    1    1    1    1    1    1    1    1    1     1
#>  [6,]    1    1    1    1    1    1    1    1    1     1
#>  [7,]    1    1    1    1    1    1    1    1    1     1
#>  [8,]    1    1    1    1    1    1    1    1    1     1
#>  [9,]    1    1    1    1    1    1    1    1    1     1
#> [10,]    1    1    1    1    1    1    1    1    1     1
#> 
#> , , 2
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    1    1    1    1    1    1    1    1    1     1
#>  [2,]    1    1    1    1    1    1    1    1    1     1
#>  [3,]    1    1    1    1    1    1    1    1    1     1
#>  [4,]    1    1    1    1    1    1    1    1    1     1
#>  [5,]    1    1    1    1    1    1    1    1    1     1
#>  [6,]    1    1    1    1    1    1    1    1    1     1
#>  [7,]    1    1    1    1    1    1    1    1    1     1
#>  [8,]    1    1    1    1    1    1    1    1    1     1
#>  [9,]    1    1    1    1    1    1    1    1    1     1
#> [10,]    1    1    1    1    1    1    1    1    1     1
#> 
#> , , 3
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    1    1    1    1    1    1    1    1    1     1
#>  [2,]    1    1    1    1    1    1    1    1    1     1
#>  [3,]    1    1    1    1    1    1    1    1    1     1
#>  [4,]    1    1    1    1    1    1    1    1    1     1
#>  [5,]    1    1    1    1    1    1    1    1    1     1
#>  [6,]    1    1    1    1    1    1    1    1    1     1
#>  [7,]    1    1    1    1    1    1    1    1    1     1
#>  [8,]    1    1    1    1    1    1    1    1    1     1
#>  [9,]    1    1    1    1    1    1    1    1    1     1
#> [10,]    1    1    1    1    1    1    1    1    1     1
#> 
#> 
#> $sconc
#> $sconc[[1]]
#> RMODFLOW 3d array with 10 rows, 10 columns and 3 layers, representing the i, j & k dimensions. 
#> Not representing stress period data 
#> , , 1
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    0    0    0    0    0    0    0    0    0     0
#>  [2,]    0    0    0    0    0    0    0    0    0     0
#>  [3,]    0    0    0    0    0    0    0    0    0     0
#>  [4,]    0    0    0    0    0    0    0    0    0     0
#>  [5,]    0    0    0    0    0    0    0    0    0     0
#>  [6,]    0    0    0    0    0    0    0    0    0     0
#>  [7,]    0    0    0    0    0    0    0    0    0     0
#>  [8,]    0    0    0    0    0    0    0    0    0     0
#>  [9,]    0    0    0    0    0    0    0    0    0     0
#> [10,]    0    0    0    0    0    0    0    0    0     0
#> 
#> , , 2
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    0    0    0    0    0    0    0    0    0     0
#>  [2,]    0    0    0    0    0    0    0    0    0     0
#>  [3,]    0    0    0    0    0    0    0    0    0     0
#>  [4,]    0    0    0    0    0    0    0    0    0     0
#>  [5,]    0    0    0    0    0    0    0    0    0     0
#>  [6,]    0    0    0    0    0    0    0    0    0     0
#>  [7,]    0    0    0    0    0    0    0    0    0     0
#>  [8,]    0    0    0    0    0    0    0    0    0     0
#>  [9,]    0    0    0    0    0    0    0    0    0     0
#> [10,]    0    0    0    0    0    0    0    0    0     0
#> 
#> , , 3
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    0    0    0    0    0    0    0    0    0     0
#>  [2,]    0    0    0    0    0    0    0    0    0     0
#>  [3,]    0    0    0    0    0    0    0    0    0     0
#>  [4,]    0    0    0    0    0    0    0    0    0     0
#>  [5,]    0    0    0    0    0    0    0    0    0     0
#>  [6,]    0    0    0    0    0    0    0    0    0     0
#>  [7,]    0    0    0    0    0    0    0    0    0     0
#>  [8,]    0    0    0    0    0    0    0    0    0     0
#>  [9,]    0    0    0    0    0    0    0    0    0     0
#> [10,]    0    0    0    0    0    0    0    0    0     0
#> 
#> 
#> 
#> $cinact
#> [1] -888
#> 
#> $thkmin
#> [1] 0.01
#> 
#> $ifmtcn
#> [1] 0
#> 
#> $ifmtnp
#> [1] 0
#> 
#> $ifmtrf
#> [1] 0
#> 
#> $ifmtdp
#> [1] 0
#> 
#> $savucn
#> [1] TRUE
#> 
#> $nprs
#> [1] -1
#> 
#> $chkmas
#> [1] TRUE
#> 
#> $nprmas
#> [1] 1
#> 
#> $perlen
#> [1] 1 1 1
#> 
#> $nstp
#> [1] 1 1 1
#> 
#> $tsmult
#> [1] 1 1 1
#> 
#> $ssflag
#> [1] FALSE FALSE FALSE
#> 
#> $dt0
#> [1] 0 0 0
#> 
#> $mxstrn
#> [1] 50 50 50
#> 
#> $ttsmult
#> [1] 1 1 1
#> 
#> $ttsmax
#> [1] 0 0 0
#> 
#> attr(,"class")
#> [1] "btn"         "rmt_package"

# multi-species; sconc for species 1 is from rnorm, sconc for species 2 = 0.2
sconc1 <- rmt_create_array(rnorm(100, 0.1, sd = 0.01), dim = c(10, 10, 3))
rmt_create_btn(ncomp = 2, prsity = prsity, sconc = list(sconc1, 0.2))

#> $modflowstylearrays
#> [1] FALSE
#> 
#> $drycell
#> [1] FALSE
#> 
#> $legacy99storage
#> [1] FALSE
#> 
#> $ftlprint
#> [1] FALSE
#> 
#> $nowetdryprint
#> [1] FALSE
#> 
#> $omitdrycellbudget
#> [1] FALSE
#> 
#> $altwtsorb
#> [1] FALSE
#> 
#> $nlay
#> [1] 3
#> 
#> $nrow
#> [1] 10
#> 
#> $ncol
#> [1] 10
#> 
#> $nper
#> [1] 1
#> 
#> $ncomp
#> [1] 2
#> 
#> $mcomp
#> [1] 2
#> 
#> $tunit
#> [1] "s"
#> 
#> $lunit
#> [1] "m"
#> 
#> $munit
#> [1] "g"
#> 
#> $trnop
#>  [1]  TRUE  TRUE  TRUE  TRUE  TRUE FALSE FALSE FALSE FALSE FALSE
#> 
#> $laycon
#> [1] 1 1 1
#> 
#> $delr
#>  [1] 100 100 100 100 100 100 100 100 100 100
#> 
#> $delc
#>  [1] 100 100 100 100 100 100 100 100 100 100
#> 
#> $htop
#> RMODFLOW 2d array with 10 rows and 10 columns, representing the i & j dimensions. 
#> Not representing stress period data 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    0    0    0    0    0    0    0    0    0     0
#>  [2,]    0    0    0    0    0    0    0    0    0     0
#>  [3,]    0    0    0    0    0    0    0    0    0     0
#>  [4,]    0    0    0    0    0    0    0    0    0     0
#>  [5,]    0    0    0    0    0    0    0    0    0     0
#>  [6,]    0    0    0    0    0    0    0    0    0     0
#>  [7,]    0    0    0    0    0    0    0    0    0     0
#>  [8,]    0    0    0    0    0    0    0    0    0     0
#>  [9,]    0    0    0    0    0    0    0    0    0     0
#> [10,]    0    0    0    0    0    0    0    0    0     0
#> 
#> $dz
#> RMODFLOW 3d array with 10 rows, 10 columns and 3 layers, representing the i, j & k dimensions. 
#> Not representing stress period data 
#> , , 1
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]   10   10   10   10   10   10   10   10   10    10
#>  [2,]   10   10   10   10   10   10   10   10   10    10
#>  [3,]   10   10   10   10   10   10   10   10   10    10
#>  [4,]   10   10   10   10   10   10   10   10   10    10
#>  [5,]   10   10   10   10   10   10   10   10   10    10
#>  [6,]   10   10   10   10   10   10   10   10   10    10
#>  [7,]   10   10   10   10   10   10   10   10   10    10
#>  [8,]   10   10   10   10   10   10   10   10   10    10
#>  [9,]   10   10   10   10   10   10   10   10   10    10
#> [10,]   10   10   10   10   10   10   10   10   10    10
#> 
#> , , 2
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]   10   10   10   10   10   10   10   10   10    10
#>  [2,]   10   10   10   10   10   10   10   10   10    10
#>  [3,]   10   10   10   10   10   10   10   10   10    10
#>  [4,]   10   10   10   10   10   10   10   10   10    10
#>  [5,]   10   10   10   10   10   10   10   10   10    10
#>  [6,]   10   10   10   10   10   10   10   10   10    10
#>  [7,]   10   10   10   10   10   10   10   10   10    10
#>  [8,]   10   10   10   10   10   10   10   10   10    10
#>  [9,]   10   10   10   10   10   10   10   10   10    10
#> [10,]   10   10   10   10   10   10   10   10   10    10
#> 
#> , , 3
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]   10   10   10   10   10   10   10   10   10    10
#>  [2,]   10   10   10   10   10   10   10   10   10    10
#>  [3,]   10   10   10   10   10   10   10   10   10    10
#>  [4,]   10   10   10   10   10   10   10   10   10    10
#>  [5,]   10   10   10   10   10   10   10   10   10    10
#>  [6,]   10   10   10   10   10   10   10   10   10    10
#>  [7,]   10   10   10   10   10   10   10   10   10    10
#>  [8,]   10   10   10   10   10   10   10   10   10    10
#>  [9,]   10   10   10   10   10   10   10   10   10    10
#> [10,]   10   10   10   10   10   10   10   10   10    10
#> 
#> 
#> $prsity
#> RMODFLOW 3d array with 10 rows, 10 columns and 3 layers, representing the i, j & k dimensions. 
#> Not representing stress period data 
#> , , 1
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1   0.1
#>  [2,]  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1   0.1
#>  [3,]  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1   0.1
#>  [4,]  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1   0.1
#>  [5,]  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1   0.1
#>  [6,]  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1   0.1
#>  [7,]  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1   0.1
#>  [8,]  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1   0.1
#>  [9,]  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1   0.1
#> [10,]  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1   0.1
#> 
#> , , 2
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01  0.01
#>  [2,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01  0.01
#>  [3,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01  0.01
#>  [4,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01  0.01
#>  [5,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01  0.01
#>  [6,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01  0.01
#>  [7,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01  0.01
#>  [8,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01  0.01
#>  [9,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01  0.01
#> [10,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01  0.01
#> 
#> , , 3
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [2,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [3,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [4,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [5,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [6,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [7,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [8,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#>  [9,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#> [10,]  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3  0.3   0.3
#> 
#> 
#> $icbund
#> RMODFLOW 3d array with 10 rows, 10 columns and 3 layers, representing the i, j & k dimensions. 
#> Not representing stress period data 
#> , , 1
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    1    1    1    1    1    1    1    1    1     1
#>  [2,]    1    1    1    1    1    1    1    1    1     1
#>  [3,]    1    1    1    1    1    1    1    1    1     1
#>  [4,]    1    1    1    1    1    1    1    1    1     1
#>  [5,]    1    1    1    1    1    1    1    1    1     1
#>  [6,]    1    1    1    1    1    1    1    1    1     1
#>  [7,]    1    1    1    1    1    1    1    1    1     1
#>  [8,]    1    1    1    1    1    1    1    1    1     1
#>  [9,]    1    1    1    1    1    1    1    1    1     1
#> [10,]    1    1    1    1    1    1    1    1    1     1
#> 
#> , , 2
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    1    1    1    1    1    1    1    1    1     1
#>  [2,]    1    1    1    1    1    1    1    1    1     1
#>  [3,]    1    1    1    1    1    1    1    1    1     1
#>  [4,]    1    1    1    1    1    1    1    1    1     1
#>  [5,]    1    1    1    1    1    1    1    1    1     1
#>  [6,]    1    1    1    1    1    1    1    1    1     1
#>  [7,]    1    1    1    1    1    1    1    1    1     1
#>  [8,]    1    1    1    1    1    1    1    1    1     1
#>  [9,]    1    1    1    1    1    1    1    1    1     1
#> [10,]    1    1    1    1    1    1    1    1    1     1
#> 
#> , , 3
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    1    1    1    1    1    1    1    1    1     1
#>  [2,]    1    1    1    1    1    1    1    1    1     1
#>  [3,]    1    1    1    1    1    1    1    1    1     1
#>  [4,]    1    1    1    1    1    1    1    1    1     1
#>  [5,]    1    1    1    1    1    1    1    1    1     1
#>  [6,]    1    1    1    1    1    1    1    1    1     1
#>  [7,]    1    1    1    1    1    1    1    1    1     1
#>  [8,]    1    1    1    1    1    1    1    1    1     1
#>  [9,]    1    1    1    1    1    1    1    1    1     1
#> [10,]    1    1    1    1    1    1    1    1    1     1
#> 
#> 
#> $sconc
#> $sconc[[1]]
#> RMODFLOW 3d array with 10 rows, 10 columns and 3 layers, representing the i, j & k dimensions. 
#> Not representing stress period data 
#> , , 1
#> 
#>             [,1]       [,2]       [,3]       [,4]       [,5]       [,6]
#>  [1,] 0.12065025 0.08695456 0.10243685 0.09950035 0.09793913 0.11113952
#>  [2,] 0.08369011 0.10737776 0.11623549 0.09748517 0.10019178 0.09754104
#>  [3,] 0.10512427 0.11888505 0.10112038 0.10444797 0.10029561 0.08822437
#>  [4,] 0.08136989 0.09902555 0.09866003 0.12755418 0.10549828 0.09024149
#>  [5,] 0.09477987 0.09064153 0.08089913 0.10046531 0.07725885 0.11065057
#>  [6,] 0.09947398 0.09984050 0.09720763 0.10577709 0.12682557 0.10131671
#>  [7,] 0.10542996 0.09173211 0.09686554 0.10118195 0.09638779 0.10488629
#>  [8,] 0.09085925 0.08487600 0.11067308 0.08088280 0.10213356 0.08300549
#>  [9,] 0.10468154 0.10935363 0.10070035 0.10862086 0.11074346 0.08529264
#> [10,] 0.10362951 0.10176489 0.09360877 0.09756763 0.09334912 0.10284150
#>             [,7]       [,8]       [,9]      [,10]
#>  [1,] 0.11337320 0.10748791 0.11048713 0.08943263
#>  [2,] 0.10236696 0.10556224 0.09961897 0.09204459
#>  [3,] 0.11318293 0.09451743 0.10486149 0.08243725
#>  [4,] 0.10523910 0.11110535 0.11672883 0.09309462
#>  [5,] 0.10606748 0.07387666 0.09645639 0.09441458
#>  [6,] 0.09890064 0.09844306 0.10946348 0.09463337
#>  [7,] 0.10172182 0.10433890 0.11316826 0.10227127
#>  [8,] 0.09909673 0.09618049 0.09703360 0.10978455
#>  [9,] 0.11924343 0.10424188 0.09612786 0.09791117
#> [10,] 0.11298393 0.11063102 0.09214567 0.08600590
#> 
#> , , 2
#> 
#>             [,1]       [,2]       [,3]       [,4]       [,5]       [,6]
#>  [1,] 0.12065025 0.08695456 0.10243685 0.09950035 0.09793913 0.11113952
#>  [2,] 0.08369011 0.10737776 0.11623549 0.09748517 0.10019178 0.09754104
#>  [3,] 0.10512427 0.11888505 0.10112038 0.10444797 0.10029561 0.08822437
#>  [4,] 0.08136989 0.09902555 0.09866003 0.12755418 0.10549828 0.09024149
#>  [5,] 0.09477987 0.09064153 0.08089913 0.10046531 0.07725885 0.11065057
#>  [6,] 0.09947398 0.09984050 0.09720763 0.10577709 0.12682557 0.10131671
#>  [7,] 0.10542996 0.09173211 0.09686554 0.10118195 0.09638779 0.10488629
#>  [8,] 0.09085925 0.08487600 0.11067308 0.08088280 0.10213356 0.08300549
#>  [9,] 0.10468154 0.10935363 0.10070035 0.10862086 0.11074346 0.08529264
#> [10,] 0.10362951 0.10176489 0.09360877 0.09756763 0.09334912 0.10284150
#>             [,7]       [,8]       [,9]      [,10]
#>  [1,] 0.11337320 0.10748791 0.11048713 0.08943263
#>  [2,] 0.10236696 0.10556224 0.09961897 0.09204459
#>  [3,] 0.11318293 0.09451743 0.10486149 0.08243725
#>  [4,] 0.10523910 0.11110535 0.11672883 0.09309462
#>  [5,] 0.10606748 0.07387666 0.09645639 0.09441458
#>  [6,] 0.09890064 0.09844306 0.10946348 0.09463337
#>  [7,] 0.10172182 0.10433890 0.11316826 0.10227127
#>  [8,] 0.09909673 0.09618049 0.09703360 0.10978455
#>  [9,] 0.11924343 0.10424188 0.09612786 0.09791117
#> [10,] 0.11298393 0.11063102 0.09214567 0.08600590
#> 
#> , , 3
#> 
#>             [,1]       [,2]       [,3]       [,4]       [,5]       [,6]
#>  [1,] 0.12065025 0.08695456 0.10243685 0.09950035 0.09793913 0.11113952
#>  [2,] 0.08369011 0.10737776 0.11623549 0.09748517 0.10019178 0.09754104
#>  [3,] 0.10512427 0.11888505 0.10112038 0.10444797 0.10029561 0.08822437
#>  [4,] 0.08136989 0.09902555 0.09866003 0.12755418 0.10549828 0.09024149
#>  [5,] 0.09477987 0.09064153 0.08089913 0.10046531 0.07725885 0.11065057
#>  [6,] 0.09947398 0.09984050 0.09720763 0.10577709 0.12682557 0.10131671
#>  [7,] 0.10542996 0.09173211 0.09686554 0.10118195 0.09638779 0.10488629
#>  [8,] 0.09085925 0.08487600 0.11067308 0.08088280 0.10213356 0.08300549
#>  [9,] 0.10468154 0.10935363 0.10070035 0.10862086 0.11074346 0.08529264
#> [10,] 0.10362951 0.10176489 0.09360877 0.09756763 0.09334912 0.10284150
#>             [,7]       [,8]       [,9]      [,10]
#>  [1,] 0.11337320 0.10748791 0.11048713 0.08943263
#>  [2,] 0.10236696 0.10556224 0.09961897 0.09204459
#>  [3,] 0.11318293 0.09451743 0.10486149 0.08243725
#>  [4,] 0.10523910 0.11110535 0.11672883 0.09309462
#>  [5,] 0.10606748 0.07387666 0.09645639 0.09441458
#>  [6,] 0.09890064 0.09844306 0.10946348 0.09463337
#>  [7,] 0.10172182 0.10433890 0.11316826 0.10227127
#>  [8,] 0.09909673 0.09618049 0.09703360 0.10978455
#>  [9,] 0.11924343 0.10424188 0.09612786 0.09791117
#> [10,] 0.11298393 0.11063102 0.09214567 0.08600590
#> 
#> 
#> $sconc[[2]]
#> RMODFLOW 3d array with 10 rows, 10 columns and 3 layers, representing the i, j & k dimensions. 
#> Not representing stress period data 
#> , , 1
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#>  [2,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#>  [3,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#>  [4,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#>  [5,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#>  [6,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#>  [7,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#>  [8,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#>  [9,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#> [10,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#> 
#> , , 2
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#>  [2,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#>  [3,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#>  [4,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#>  [5,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#>  [6,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#>  [7,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#>  [8,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#>  [9,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#> [10,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#> 
#> , , 3
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#>  [2,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#>  [3,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#>  [4,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#>  [5,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#>  [6,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#>  [7,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#>  [8,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#>  [9,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#> [10,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#> 
#> 
#> 
#> $cinact
#> [1] -888
#> 
#> $thkmin
#> [1] 0.01
#> 
#> $ifmtcn
#> [1] 0
#> 
#> $ifmtnp
#> [1] 0
#> 
#> $ifmtrf
#> [1] 0
#> 
#> $ifmtdp
#> [1] 0
#> 
#> $savucn
#> [1] TRUE
#> 
#> $nprs
#> [1] -1
#> 
#> $chkmas
#> [1] TRUE
#> 
#> $nprmas
#> [1] 1
#> 
#> $perlen
#> [1] 1
#> 
#> $nstp
#> [1] 1
#> 
#> $tsmult
#> [1] 1
#> 
#> $ssflag
#> [1] FALSE
#> 
#> $dt0
#> [1] 0
#> 
#> $mxstrn
#> [1] 50
#> 
#> $ttsmult
#> [1] 1
#> 
#> $ttsmax
#> [1] 0
#> 
#> attr(,"class")
#> [1] "btn"         "rmt_package"