Create an RMT3DMS dsp object

rmt_create_dsp creates an RMT3DMS dsp object.

rmt_create_dsp(
  btn,
  al = 10,
  trpt = 0.1,
  trpv = 0.01,
  dmcoef = 0,
  multidiffusion = FALSE,
  nocross = FALSE
)

Arguments

btn
RMT3DMS btn object

al
3D array with longitudinal dispersivities. Defaults to 10 for every cell.

trpt
1D vector with ratios of horizontal transverse to longitudinal dispersivities (TH/TL) for every layer. Defaults to 0.1 for every layer.

trpv
1D vector with ratios of vertical transverse to longitudinal dispersivities (TV/TL) for every layer. Defaults to 0.01 for every layer.

dmcoef
either a 1D (nlay) (multidiffusion = FALSE) or a list of length btn$mcomp with 3D (multidiffusion = TRUE) array with the effective molecular diffusion coefficients. Defaults to 0 for all layers or cells (i.e. diffusion is neglected).

multidiffusion
logical; if TRUE, dmcoef can be specified for each species, for each cell. Defaults to FALSE.

nocross
logical; should cross-dispersion be disabled? Defaults to FALSE. MT3D-USGS only.

Value

an object of class dsp

See also

Examples

btn <- rmt_create_btn(ncomp = 3)
rmt_create_dsp(btn)

#> $multidiffusion
#> [1] FALSE
#> 
#> $nocross
#> [1] FALSE
#> 
#> $al
#> RMODFLOW 3d array with 10 rows, 10 columns and 3 layers, representing the i, j & k dimensions. 
#> Not representing stress period data 
#> , , 1
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]   10   10   10   10   10   10   10   10   10    10
#>  [2,]   10   10   10   10   10   10   10   10   10    10
#>  [3,]   10   10   10   10   10   10   10   10   10    10
#>  [4,]   10   10   10   10   10   10   10   10   10    10
#>  [5,]   10   10   10   10   10   10   10   10   10    10
#>  [6,]   10   10   10   10   10   10   10   10   10    10
#>  [7,]   10   10   10   10   10   10   10   10   10    10
#>  [8,]   10   10   10   10   10   10   10   10   10    10
#>  [9,]   10   10   10   10   10   10   10   10   10    10
#> [10,]   10   10   10   10   10   10   10   10   10    10
#> 
#> , , 2
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]   10   10   10   10   10   10   10   10   10    10
#>  [2,]   10   10   10   10   10   10   10   10   10    10
#>  [3,]   10   10   10   10   10   10   10   10   10    10
#>  [4,]   10   10   10   10   10   10   10   10   10    10
#>  [5,]   10   10   10   10   10   10   10   10   10    10
#>  [6,]   10   10   10   10   10   10   10   10   10    10
#>  [7,]   10   10   10   10   10   10   10   10   10    10
#>  [8,]   10   10   10   10   10   10   10   10   10    10
#>  [9,]   10   10   10   10   10   10   10   10   10    10
#> [10,]   10   10   10   10   10   10   10   10   10    10
#> 
#> , , 3
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]   10   10   10   10   10   10   10   10   10    10
#>  [2,]   10   10   10   10   10   10   10   10   10    10
#>  [3,]   10   10   10   10   10   10   10   10   10    10
#>  [4,]   10   10   10   10   10   10   10   10   10    10
#>  [5,]   10   10   10   10   10   10   10   10   10    10
#>  [6,]   10   10   10   10   10   10   10   10   10    10
#>  [7,]   10   10   10   10   10   10   10   10   10    10
#>  [8,]   10   10   10   10   10   10   10   10   10    10
#>  [9,]   10   10   10   10   10   10   10   10   10    10
#> [10,]   10   10   10   10   10   10   10   10   10    10
#> 
#> 
#> $trpt
#> [1] 0.1 0.1 0.1
#> 
#> $trpv
#> [1] 0.01 0.01 0.01
#> 
#> $dmcoef
#> [1] 0 0 0
#> 
#> attr(,"class")
#> [1] "dsp"         "rmt_package"

rmt_create_dsp(btn, dmcoef = list(0, 0.1, 0.2), multidiffusion = TRUE)

#> $multidiffusion
#> [1] TRUE
#> 
#> $nocross
#> [1] FALSE
#> 
#> $al
#> RMODFLOW 3d array with 10 rows, 10 columns and 3 layers, representing the i, j & k dimensions. 
#> Not representing stress period data 
#> , , 1
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]   10   10   10   10   10   10   10   10   10    10
#>  [2,]   10   10   10   10   10   10   10   10   10    10
#>  [3,]   10   10   10   10   10   10   10   10   10    10
#>  [4,]   10   10   10   10   10   10   10   10   10    10
#>  [5,]   10   10   10   10   10   10   10   10   10    10
#>  [6,]   10   10   10   10   10   10   10   10   10    10
#>  [7,]   10   10   10   10   10   10   10   10   10    10
#>  [8,]   10   10   10   10   10   10   10   10   10    10
#>  [9,]   10   10   10   10   10   10   10   10   10    10
#> [10,]   10   10   10   10   10   10   10   10   10    10
#> 
#> , , 2
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]   10   10   10   10   10   10   10   10   10    10
#>  [2,]   10   10   10   10   10   10   10   10   10    10
#>  [3,]   10   10   10   10   10   10   10   10   10    10
#>  [4,]   10   10   10   10   10   10   10   10   10    10
#>  [5,]   10   10   10   10   10   10   10   10   10    10
#>  [6,]   10   10   10   10   10   10   10   10   10    10
#>  [7,]   10   10   10   10   10   10   10   10   10    10
#>  [8,]   10   10   10   10   10   10   10   10   10    10
#>  [9,]   10   10   10   10   10   10   10   10   10    10
#> [10,]   10   10   10   10   10   10   10   10   10    10
#> 
#> , , 3
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]   10   10   10   10   10   10   10   10   10    10
#>  [2,]   10   10   10   10   10   10   10   10   10    10
#>  [3,]   10   10   10   10   10   10   10   10   10    10
#>  [4,]   10   10   10   10   10   10   10   10   10    10
#>  [5,]   10   10   10   10   10   10   10   10   10    10
#>  [6,]   10   10   10   10   10   10   10   10   10    10
#>  [7,]   10   10   10   10   10   10   10   10   10    10
#>  [8,]   10   10   10   10   10   10   10   10   10    10
#>  [9,]   10   10   10   10   10   10   10   10   10    10
#> [10,]   10   10   10   10   10   10   10   10   10    10
#> 
#> 
#> $trpt
#> [1] 0.1 0.1 0.1
#> 
#> $trpv
#> [1] 0.01 0.01 0.01
#> 
#> $dmcoef
#> $dmcoef[[1]]
#> RMODFLOW 3d array with 10 rows, 10 columns and 3 layers, representing the i, j & k dimensions. 
#> Not representing stress period data 
#> , , 1
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    0    0    0    0    0    0    0    0    0     0
#>  [2,]    0    0    0    0    0    0    0    0    0     0
#>  [3,]    0    0    0    0    0    0    0    0    0     0
#>  [4,]    0    0    0    0    0    0    0    0    0     0
#>  [5,]    0    0    0    0    0    0    0    0    0     0
#>  [6,]    0    0    0    0    0    0    0    0    0     0
#>  [7,]    0    0    0    0    0    0    0    0    0     0
#>  [8,]    0    0    0    0    0    0    0    0    0     0
#>  [9,]    0    0    0    0    0    0    0    0    0     0
#> [10,]    0    0    0    0    0    0    0    0    0     0
#> 
#> , , 2
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    0    0    0    0    0    0    0    0    0     0
#>  [2,]    0    0    0    0    0    0    0    0    0     0
#>  [3,]    0    0    0    0    0    0    0    0    0     0
#>  [4,]    0    0    0    0    0    0    0    0    0     0
#>  [5,]    0    0    0    0    0    0    0    0    0     0
#>  [6,]    0    0    0    0    0    0    0    0    0     0
#>  [7,]    0    0    0    0    0    0    0    0    0     0
#>  [8,]    0    0    0    0    0    0    0    0    0     0
#>  [9,]    0    0    0    0    0    0    0    0    0     0
#> [10,]    0    0    0    0    0    0    0    0    0     0
#> 
#> , , 3
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    0    0    0    0    0    0    0    0    0     0
#>  [2,]    0    0    0    0    0    0    0    0    0     0
#>  [3,]    0    0    0    0    0    0    0    0    0     0
#>  [4,]    0    0    0    0    0    0    0    0    0     0
#>  [5,]    0    0    0    0    0    0    0    0    0     0
#>  [6,]    0    0    0    0    0    0    0    0    0     0
#>  [7,]    0    0    0    0    0    0    0    0    0     0
#>  [8,]    0    0    0    0    0    0    0    0    0     0
#>  [9,]    0    0    0    0    0    0    0    0    0     0
#> [10,]    0    0    0    0    0    0    0    0    0     0
#> 
#> 
#> $dmcoef[[2]]
#> RMODFLOW 3d array with 10 rows, 10 columns and 3 layers, representing the i, j & k dimensions. 
#> Not representing stress period data 
#> , , 1
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1   0.1
#>  [2,]  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1   0.1
#>  [3,]  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1   0.1
#>  [4,]  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1   0.1
#>  [5,]  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1   0.1
#>  [6,]  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1   0.1
#>  [7,]  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1   0.1
#>  [8,]  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1   0.1
#>  [9,]  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1   0.1
#> [10,]  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1   0.1
#> 
#> , , 2
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1   0.1
#>  [2,]  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1   0.1
#>  [3,]  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1   0.1
#>  [4,]  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1   0.1
#>  [5,]  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1   0.1
#>  [6,]  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1   0.1
#>  [7,]  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1   0.1
#>  [8,]  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1   0.1
#>  [9,]  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1   0.1
#> [10,]  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1   0.1
#> 
#> , , 3
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1   0.1
#>  [2,]  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1   0.1
#>  [3,]  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1   0.1
#>  [4,]  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1   0.1
#>  [5,]  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1   0.1
#>  [6,]  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1   0.1
#>  [7,]  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1   0.1
#>  [8,]  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1   0.1
#>  [9,]  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1   0.1
#> [10,]  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1   0.1
#> 
#> 
#> $dmcoef[[3]]
#> RMODFLOW 3d array with 10 rows, 10 columns and 3 layers, representing the i, j & k dimensions. 
#> Not representing stress period data 
#> , , 1
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#>  [2,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#>  [3,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#>  [4,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#>  [5,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#>  [6,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#>  [7,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#>  [8,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#>  [9,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#> [10,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#> 
#> , , 2
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#>  [2,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#>  [3,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#>  [4,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#>  [5,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#>  [6,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#>  [7,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#>  [8,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#>  [9,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#> [10,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#> 
#> , , 3
#> 
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#>  [2,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#>  [3,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#>  [4,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#>  [5,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#>  [6,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#>  [7,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#>  [8,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#>  [9,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#> [10,]  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2  0.2   0.2
#> 
#> 
#> 
#> attr(,"class")
#> [1] "dsp"         "rmt_package"